REVERT to older implementation of fan curve write optimization
Newer solution was not behaving as expected, especially upon Windows startup. CPU Fan would spin up to it's highest speed and FanControl would not be able to control the fan for 30-60 seconds after. Creating an array that stores the LAST REQUESTED speed and comparing it to a newly requested speed appears to be the culprit. We were never actually checking the target fan's speed. We were simply comparing the last known requested PWM Signal strength request vs the new one. Eg. The CPU fan is spinning at 100%, and we requested 80%, and a request comes in to set it to 80% again, the condition exit's the function early, since the speed we had previously requested now matches the current request, regardless of the CPU fan still spinning at the wrong speed. This can lead to situations where a fan speed change quest is sent once, and it it's not obeyed immediately, the change never happens.